ejb4/web/src/main/java/com/github/aushacker/web/BonjourServlet.java
package com.github.aushacker.web;

import java.io.IOException;

import jakarta.servlet.annotation.WebServlet;
import jakarta.servlet.http.HttpServlet;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;

/**
 * Newer style servlet using annotations.
 * 
 * @since 1.0
 */
@WebServlet("/bonjour")
public class BonjourServlet extends HttpServlet {

    @Override
    public void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
        resp.setStatus(HttpServletResponse.SC_OK);
        resp.setContentType("text/plain");
        resp.getWriter().println("Bonjour Monde!");
    }

}
